pyproject.toml
[project]
name = "py-instructor"
version = "0.1.0"
description = "Add your description here"
authors = [
    { name = "jlondonobo", email = "j.londonobo@gmail.com" }
]
dependencies = [
    "instructor>=0.6.1",
    "python-dotenv>=1.0.1",
]
readme = "README.md"
requires-python = ">= 3.10"

[project.scripts]
hello = "py_instructor:hello"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.rye]
managed = true
dev-dependencies = [
    "ipykernel>=6.29.2",
]

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build.targets.wheel]
packages = ["src/py_instructor"]
